% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extractProcess}
\alias{extractProcess}
\title{Capturar processos}
\usage{
extractProcess(categoria, min = 1)
}
\arguments{
\item{categoria}{O tipo de ação: ADIN, ADPF, ADCN ou ADOM}

\item{min}{Tamanho mínimo da série de processos coletada. min = 1 é o default, com isso coleta todas as ações
de uma categoria}
}
\value{
Um dataframe com os meta-dados das ações
}
\description{
Esta função captura os meta-dados de todas as ações de controle concentrado
do Supremo Tribunal Federal (STF).
}
