% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extractProcess}
\alias{extractProcess}
\title{Capturar processos}
\usage{
extractProcess(categoria = cat, min = 1)
}
\arguments{
\item{infile}{O tipo de ação e o intervalo mínimo}
}
\value{
Um dataframe com os meta-dados das ações
}
\description{
Esta função captura os meta-dados de todas as ações de controle concentrado
do Supremo Tribunal Federal (STF). Para isso, o usuário precisa especificar
apenas a categoria da ação (ADIN, ADPF, ADCN ou ADOM) e o tamanho mínimo da série
de processos que ele deseja (min = 1 é o default, com isso coleta todas as ações
uma categoria)
}
